remote, util: Move general helper routines from `remote.py` to `util.py`
